update reference to data.table course on DataCamp (#12)
The link provided returns an archived course, according to the site. The site reads, "This course is exclusively available to customers. This course has been archived. If you are taking this course you will be able to finish it. If you are working on a track containing this course it will remain available until you complete the track."

However, another data.table course, from the same author, is available at https://www.datacamp.com/courses/data-manipulation-with-datatable-in-r.

archived:  https://www.datacamp.com/courses/data-table-data-manipulation-r-tutorial
available: https://www.datacamp.com/courses/data-manipulation-with-datatable-in-r